requirements.txt
Django>=5.2.1
celery>=5.3.0
Pillow>=10.0.0
redis>=5.0.0
Faker>=20.0.0
python-magic>=0.4.27 